Refactor model info processing
Most of this is just to reduce duplicated code, but I got a bit
carried away. I also ended up sanitizing HTML coming in from
civitai and refactoring the recently added code that checks for
existing preview images. The extension now downloads metadata
for sd webui if it is missing, even if civitai metadata already
exists.

I'd like to work on reducing hits to civitai's servers, but it's
a work in progress.