Do something about https://github.com/zixaphir/Stable-Diffusion-Webui-Civitai-Helper/issues/10
This is not ideal but given the ideal solution isn't going to
happen (which is: don't parse the description as HTML), this is
the best middle-ground I can come up with. Mind you, it's not
fool-proof: if for some reason a model author on civitai puts
`<!-- -->` in their model description, it will break, but it
it keeps virtually all models from breaking webui otherwise,
and allows the user to remove them if they want to restore
the description's display within webui's model cards.