Support DoRA models as extensions of LoRA
It is currently my belief that the technical differences between
extra network models does not matter to the majority of users. As
long as their setup supports them, there is functionally no reason
to try and sort them into completely different directories. At
this point, lyco models are only in a separate folder because of
historical debt, but I'd go further and say that even models such
as hypernetworks and textual embeddings having their own folders
is silly. The end user does not and should not have to care about
what a Lora does versus a TE: all they care about is that it
accomplishes the task they have designated for it.

However, the current implementations are what they are, and thus
the end user expects things to be in certain places. However,
moving forward, any evolutions or extensions to LoRA will just be
dumped into the LoRA folder. I do not care what they're named, I
do not care what makes them different or special. As long as they
accomplish the goal of manipulating a checkpoint model to perform
an a task either it couldn't before or to do it differently,
whether better or worse, it is just a LoRA to me, and developers
further up the chain can work out how their implementations handle
actually using them.