Move all NSFW Preview options to Settings
The Browser will maintain its own NSFW filter *for models* but will
also use the NSFW Preview filter for previews.

Honestly, this will eliminate two issues:
1) The NSFW Preview filter has caused issues with Vladmandic
SD-Next that I have no idea how to fix as it was previously
implemented. This works around that issue.
2) The NSFW Preview filter being different for every section of
Civitai Helper was horrid. Yes, now you have to navigate to the
settings to change it, but you should only have to change it once
and then everything will behave the same.

The Browser behaving differently is honestly fine, because the
NSFW Filter in the Browser only filters the models themselves. It
is entirely possible a user would like to search for NSFW models
without littering their screen with a bunch of NSFW imagery.
Furthermore, seemingly SFW models can also have NSFW previews.